refactor user api
